Refactor runtime logic into dedicated files
