+++ internal/core/runtime/runtime.go
@@ -1,19 +1,9 @@
 package runtime
 
 import (
-	"bufio"
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"os"
-	"sort"
 	"strings"
 	"sync"
 	"time"
-
-	"github.com/asynkron/goagent/internal/core/schema"
-	"github.com/xeipuuv/gojsonschema"
 )
 
 // Runtime is the Go counterpart to the TypeScript AgentRuntime. It exposes two
@@ -41,12 +31,6 @@ type Runtime struct {
 	historyMu sync.RWMutex
 	history   []ChatMessage
 }
-
-var (
-	planSchemaLoader     gojsonschema.JSONLoader
-	planSchemaLoaderErr  error
-	planSchemaLoaderOnce sync.Once
-)
 
 // NewRuntime configures a new runtime with the provided options.
 func NewRuntime(options RuntimeOptions) (*Runtime, error) {
@@ -132,849 +116,6 @@ func (r *Runtime) enqueue(evt InputEvent) {
 	}
 }
 
-// Run starts the runtime loop and optionally bridges stdin/stdout to the
-// respective channels so the binary is immediately useful in a terminal.
-func (r *Runtime) Run(ctx context.Context) error {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	var wg sync.WaitGroup
-
-	if !r.options.DisableOutputForwarding {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			r.forwardOutputs(ctx)
-		}()
-	}
-
-	if !r.options.DisableInputReader {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			if err := r.consumeInput(ctx); err != nil {
-				r.emit(RuntimeEvent{
-					Type:    EventTypeError,
-					Message: err.Error(),
-					Level:   StatusLevelError,
-				})
-			}
-		}()
-	}
-
-	err := r.loop(ctx)
-	cancel()
-	wg.Wait()
-
-	return err
-}
-
-func (r *Runtime) loop(ctx context.Context) error {
-	r.emit(RuntimeEvent{
-		Type:    EventTypeStatus,
-		Message: "Agent runtime started",
-		Level:   StatusLevelInfo,
-	})
-	r.emit(RuntimeEvent{
-		Type:    EventTypeRequestInput,
-		Message: "Enter a prompt to begin.",
-		Level:   StatusLevelInfo,
-	})
-
-	for {
-		select {
-		case <-ctx.Done():
-			r.emit(RuntimeEvent{
-				Type:    EventTypeStatus,
-				Message: "Context cancelled. Shutting down runtime.",
-				Level:   StatusLevelWarn,
-			})
-			r.close()
-			return ctx.Err()
-		case evt, ok := <-r.inputs:
-			if !ok {
-				r.close()
-				return nil
-			}
-			if err := r.handleInput(ctx, evt); err != nil {
-				r.emit(RuntimeEvent{
-					Type:    EventTypeError,
-					Message: err.Error(),
-					Level:   StatusLevelError,
-				})
-				r.close()
-				return err
-			}
-		}
-	}
-}
-
-func (r *Runtime) handleInput(ctx context.Context, evt InputEvent) error {
-	switch evt.Type {
-	case InputTypePrompt:
-		return r.handlePrompt(ctx, evt)
-	case InputTypeCancel:
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: fmt.Sprintf("Cancel requested: %s", strings.TrimSpace(evt.Reason)),
-			Level:   StatusLevelWarn,
-		})
-		r.emit(RuntimeEvent{
-			Type:    EventTypeRequestInput,
-			Message: "Ready for the next instruction.",
-			Level:   StatusLevelInfo,
-		})
-		return nil
-	case InputTypeShutdown:
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: "Shutdown requested. Goodbye!",
-			Level:   StatusLevelInfo,
-		})
-		r.close()
-		return errors.New("runtime shutdown requested")
-	default:
-		return fmt.Errorf("unknown input type: %s", evt.Type)
-	}
-}
-
-func (r *Runtime) handlePrompt(ctx context.Context, evt InputEvent) error {
-	prompt := strings.TrimSpace(evt.Prompt)
-	if prompt == "" {
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: "Ignoring empty prompt.",
-			Level:   StatusLevelWarn,
-		})
-		r.emit(RuntimeEvent{
-			Type:    EventTypeRequestInput,
-			Message: "Awaiting a non-empty prompt.",
-			Level:   StatusLevelInfo,
-		})
-		return nil
-	}
-
-	if !r.beginWork() {
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: "Agent is already processing another prompt.",
-			Level:   StatusLevelWarn,
-		})
-		return nil
-	}
-	defer r.endWork()
-
-	r.emit(RuntimeEvent{
-		Type:    EventTypeStatus,
-		Message: fmt.Sprintf("Processing prompt with model %s…", r.options.Model),
-		Level:   StatusLevelInfo,
-	})
-
-	userMessage := ChatMessage{Role: RoleUser, Content: prompt, Timestamp: time.Now()}
-	r.appendHistory(userMessage)
-
-	r.planExecutionLoop(ctx)
-
-	return nil
-}
-
-func (r *Runtime) planExecutionLoop(ctx context.Context) {
-	for {
-		if ctx.Err() != nil {
-			return
-		}
-
-		plan, toolCall, err := r.requestPlan(ctx)
-		if err != nil {
-			r.emit(RuntimeEvent{
-				Type:    EventTypeError,
-				Message: fmt.Sprintf("Failed to contact OpenAI: %v", err),
-				Level:   StatusLevelError,
-			})
-			r.emit(RuntimeEvent{
-				Type:    EventTypeRequestInput,
-				Message: "You can provide another prompt.",
-				Level:   StatusLevelInfo,
-			})
-			return
-		}
-
-		if plan == nil {
-			r.emit(RuntimeEvent{
-				Type:    EventTypeError,
-				Message: "Received nil plan response.",
-				Level:   StatusLevelError,
-			})
-			r.emit(RuntimeEvent{
-				Type:    EventTypeRequestInput,
-				Message: "Unable to continue plan execution. Provide the next instruction.",
-				Level:   StatusLevelInfo,
-			})
-			return
-		}
-
-		execCount := r.recordPlanResponse(plan, toolCall)
-
-		if plan.RequireHumanInput {
-			// The assistant explicitly requested help from the human, so surface the
-			// request and pause automated execution until the user responds.
-			r.appendToolObservation(toolCall, PlanObservationPayload{
-				Summary: "Assistant requested additional input before continuing the plan.",
-			})
-			r.emit(RuntimeEvent{
-				Type:    EventTypeRequestInput,
-				Message: "Assistant requested additional input before continuing.",
-				Level:   StatusLevelInfo,
-			})
-			return
-		}
-
-		if execCount == 0 {
-			r.appendToolObservation(toolCall, PlanObservationPayload{
-				Summary: "Assistant returned a plan without executable steps.",
-			})
-			r.emit(RuntimeEvent{
-				Type:    EventTypeStatus,
-				Message: "Plan has no executable steps.",
-				Level:   StatusLevelInfo,
-			})
-			r.emit(RuntimeEvent{
-				Type:    EventTypeRequestInput,
-				Message: "Plan has no executable steps. Provide the next instruction.",
-				Level:   StatusLevelInfo,
-			})
-			return
-		}
-
-		r.executePendingCommands(ctx, toolCall)
-		if ctx.Err() != nil {
-			return
-		}
-	}
-}
-
-// requestPlan centralizes the logic for requesting a new plan from the assistant.
-// It snapshots the history to guarantee a consistent view, forwards the request
-// to the OpenAI client, and emits a status update so hosts can surface that a
-// response was received.
-func (r *Runtime) requestPlan(ctx context.Context) (*PlanResponse, ToolCall, error) {
-	var retryCount int
-	for {
-		history := r.historySnapshot()
-
-		r.writeHistoryLog(history)
-
-		toolCall, err := r.client.RequestPlan(ctx, history)
-		if err != nil {
-			return nil, ToolCall{}, err
-		}
-
-		plan, retry, validationErr := r.validatePlanToolCall(toolCall)
-		if validationErr != nil {
-			return nil, ToolCall{}, validationErr
-		}
-		if retry {
-			retryCount++
-			delay := computeValidationBackoff(retryCount)
-			select {
-			case <-time.After(delay):
-			case <-ctx.Done():
-				return nil, ToolCall{}, ctx.Err()
-			}
-			continue
-		}
-
-		retryCount = 0
-
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: "Assistant response received.",
-			Level:   StatusLevelInfo,
-		})
-
-		return plan, toolCall, nil
-	}
-}
-
-const (
-	validationDetailLimit   = 512
-	validationBackoffBase   = 250 * time.Millisecond
-	validationBackoffMax    = 4 * time.Second
-	validationBackoffMaxExp = 5
-)
-
-type schemaValidationError struct {
-	issues []string
-}
-
-func (e schemaValidationError) Error() string {
-	if len(e.issues) == 0 {
-		return "plan response failed schema validation"
-	}
-	return strings.Join(e.issues, "; ")
-}
-
-// validatePlanToolCall ensures the assistant response is valid JSON and
-// satisfies the plan schema before we hydrate a PlanResponse structure.
-// Returning retry=true signals that the helper produced feedback for the
-// assistant and the runtime should request a new plan immediately.
-func (r *Runtime) validatePlanToolCall(toolCall ToolCall) (*PlanResponse, bool, error) {
-	trimmedArgs := strings.TrimSpace(toolCall.Arguments)
-	if trimmedArgs == "" {
-		payload := PlanObservationPayload{
-			JSONParseError:          true,
-			ResponseValidationError: true,
-			Summary:                 "Assistant called the tool without providing arguments.",
-			Details:                 "tool arguments were empty",
-		}
-		r.handlePlanValidationFailure(toolCall, payload, r.buildValidationAutoPrompt(payload))
-		return nil, true, nil
-	}
-
-	var plan PlanResponse
-	if err := json.Unmarshal([]byte(toolCall.Arguments), &plan); err != nil {
-		payload := PlanObservationPayload{
-			JSONParseError:          true,
-			ResponseValidationError: true,
-			Summary:                 "Tool call arguments were not valid JSON.",
-			Details:                 err.Error(),
-		}
-		r.handlePlanValidationFailure(toolCall, payload, r.buildValidationAutoPrompt(payload))
-		return nil, true, nil
-	}
-
-	if err := validatePlanAgainstSchema(toolCall.Arguments); err != nil {
-		var schemaErr schemaValidationError
-		if errors.As(err, &schemaErr) {
-			payload := PlanObservationPayload{
-				SchemaValidationError:   true,
-				ResponseValidationError: true,
-				Summary:                 "Tool call arguments failed schema validation.",
-				Details:                 schemaErr.Error(),
-			}
-			r.handlePlanValidationFailure(toolCall, payload, r.buildValidationAutoPrompt(payload))
-			return nil, true, nil
-		}
-		return nil, false, err
-	}
-
-	return &plan, false, nil
-}
-
-func validatePlanAgainstSchema(raw string) error {
-	loader, err := loadPlanSchema()
-	if err != nil {
-		return fmt.Errorf("runtime: load plan schema: %w", err)
-	}
-
-	result, err := gojsonschema.Validate(loader, gojsonschema.NewStringLoader(raw))
-	if err != nil {
-		return fmt.Errorf("runtime: schema validation error: %w", err)
-	}
-	if result.Valid() {
-		return nil
-	}
-
-	issues := make([]string, 0, len(result.Errors()))
-	for _, desc := range result.Errors() {
-		issues = append(issues, desc.String())
-	}
-	return schemaValidationError{issues: issues}
-}
-
-func loadPlanSchema() (gojsonschema.JSONLoader, error) {
-	planSchemaLoaderOnce.Do(func() {
-		schemaMap, err := schema.PlanResponseSchema()
-		if err != nil {
-			planSchemaLoaderErr = err
-			return
-		}
-		planSchemaLoader = gojsonschema.NewGoLoader(schemaMap)
-	})
-	if planSchemaLoaderErr != nil {
-		return nil, planSchemaLoaderErr
-	}
-	return planSchemaLoader, nil
-}
-
-func (r *Runtime) handlePlanValidationFailure(toolCall ToolCall, payload PlanObservationPayload, autoPrompt string) {
-	payload.Details = strings.TrimSpace(payload.Details)
-
-	metadata := map[string]any{
-		"details": payload.Details,
-	}
-	if toolCall.ID != "" {
-		metadata["tool_call_id"] = toolCall.ID
-	}
-	if toolCall.Name != "" {
-		metadata["tool_name"] = toolCall.Name
-	}
-
-	message := payload.Summary
-	if details := strings.TrimSpace(payload.Details); details != "" {
-		message = fmt.Sprintf("%s Details: %s", message, details)
-	}
-
-	r.emit(RuntimeEvent{
-		Type:     EventTypeStatus,
-		Message:  message,
-		Level:    StatusLevelWarn,
-		Metadata: metadata,
-	})
-
-	r.appendHistory(ChatMessage{
-		Role:      RoleAssistant,
-		Timestamp: time.Now(),
-		ToolCalls: []ToolCall{toolCall},
-	})
-
-	if toolCall.ID != "" {
-		if toolMessage, err := BuildToolMessage(payload); err != nil {
-			r.emit(RuntimeEvent{
-				Type:    EventTypeError,
-				Message: fmt.Sprintf("Failed to encode validation feedback: %v", err),
-				Level:   StatusLevelError,
-			})
-		} else {
-			r.appendHistory(ChatMessage{
-				Role:       RoleTool,
-				Content:    toolMessage,
-				ToolCallID: toolCall.ID,
-				Name:       toolCall.Name,
-				Timestamp:  time.Now(),
-			})
-		}
-	}
-
-	if strings.TrimSpace(autoPrompt) != "" {
-		r.appendHistory(ChatMessage{
-			Role:      RoleUser,
-			Content:   autoPrompt,
-			Timestamp: time.Now(),
-		})
-	}
-}
-
-func (r *Runtime) buildValidationAutoPrompt(payload PlanObservationPayload) string {
-	summary := strings.TrimSpace(payload.Summary)
-	if summary == "" {
-		summary = "The previous tool call response could not be processed."
-	}
-	details := truncateForPrompt(strings.TrimSpace(payload.Details), validationDetailLimit)
-
-	builder := strings.Builder{}
-	builder.WriteString(summary)
-	if details != "" {
-		builder.WriteString(" Details: ")
-		builder.WriteString(details)
-	}
-	builder.WriteString(" Please call ")
-	builder.WriteString(schema.ToolName)
-	builder.WriteString(" again with JSON that strictly matches the provided schema.")
-	return builder.String()
-}
-
-func truncateForPrompt(value string, limit int) string {
-	if limit <= 0 {
-		return value
-	}
-	runes := []rune(value)
-	if len(runes) <= limit {
-		return value
-	}
-	return string(runes[:limit]) + "…"
-}
-
-func computeValidationBackoff(attempt int) time.Duration {
-	if attempt <= 0 {
-		attempt = 1
-	}
-	exp := attempt - 1
-	if exp > validationBackoffMaxExp {
-		exp = validationBackoffMaxExp
-	}
-
-	multiplier := 1 << exp
-	delay := validationBackoffBase * time.Duration(multiplier)
-	if delay > validationBackoffMax {
-		return validationBackoffMax
-	}
-	if delay < validationBackoffBase {
-		return validationBackoffBase
-	}
-	return delay
-}
-
-// filterCompletedSteps drops plan entries that already finished so the executor
-// does not revisit completed work when the model resubmits the same plan. Any
-// dependency edges pointing at removed steps are also pruned so remaining steps
-// no longer wait on work that has already finished.
-func filterCompletedSteps(steps []PlanStep) []PlanStep {
-	if len(steps) == 0 {
-		return steps
-	}
-
-	completedIDs := make(map[string]struct{})
-	filtered := make([]PlanStep, 0, len(steps))
-	for _, step := range steps {
-		if step.Status == PlanCompleted {
-			completedIDs[step.ID] = struct{}{}
-			continue
-		}
-		filtered = append(filtered, step)
-	}
-
-	if len(completedIDs) == 0 {
-		return filtered
-	}
-
-	for i := range filtered {
-		deps := filtered[i].WaitingForID
-		if len(deps) == 0 {
-			continue
-		}
-
-		trimNeeded := false
-		for _, dep := range deps {
-			if _, done := completedIDs[dep]; done {
-				trimNeeded = true
-				break
-			}
-		}
-		if !trimNeeded {
-			continue
-		}
-
-		pruned := make([]string, 0, len(deps))
-		for _, dep := range deps {
-			if _, done := completedIDs[dep]; done {
-				continue
-			}
-			pruned = append(pruned, dep)
-		}
-
-		if len(pruned) == 0 {
-			filtered[i].WaitingForID = nil
-			continue
-		}
-
-		filtered[i].WaitingForID = pruned
-	}
-
-	return filtered
-}
-
-func (r *Runtime) recordPlanResponse(plan *PlanResponse, toolCall ToolCall) int {
-	assistantMessage := ChatMessage{
-		Role:      RoleAssistant,
-		Timestamp: time.Now(),
-		ToolCalls: []ToolCall{toolCall},
-	}
-	r.appendHistory(assistantMessage)
-
-	trimmedPlan := filterCompletedSteps(plan.Plan)
-	r.plan.Replace(trimmedPlan)
-
-	planMetadata := map[string]any{
-		"plan":                trimmedPlan,
-		"tool_call_id":        toolCall.ID,
-		"tool_name":           toolCall.Name,
-		"require_human_input": plan.RequireHumanInput,
-	}
-	if strings.TrimSpace(plan.Reasoning) != "" {
-		planMetadata["reasoning"] = plan.Reasoning
-	}
-
-	r.emit(RuntimeEvent{
-		Type:    EventTypeStatus,
-		Message: fmt.Sprintf("Received plan with %d step(s).", len(trimmedPlan)),
-		Level:   StatusLevelInfo,
-		Metadata: map[string]any{
-			"tool_call_id": toolCall.ID,
-		},
-	})
-
-	r.emit(RuntimeEvent{
-		Type:     EventTypeAssistantMessage,
-		Message:  plan.Message,
-		Level:    StatusLevelInfo,
-		Metadata: planMetadata,
-	})
-
-	return r.plan.ExecutableCount()
-}
-
-func (r *Runtime) executePendingCommands(ctx context.Context, toolCall ToolCall) {
-	r.commandMu.Lock()
-	defer r.commandMu.Unlock()
-
-	var (
-		executedSteps   int
-		lastStepID      string
-		lastObservation PlanObservationPayload
-		haveObservation bool
-		finalErr        error
-	)
-
-	stepResults := make(map[string]StepObservation)
-
-	for {
-		if ctx.Err() != nil {
-			finalErr = ctx.Err()
-			break
-		}
-
-		step, ok := r.plan.Ready()
-		if !ok {
-			if !r.plan.HasPending() {
-				r.emit(RuntimeEvent{
-					Type:    EventTypeStatus,
-					Message: "Plan execution completed.",
-					Level:   StatusLevelInfo,
-				})
-			}
-			break
-		}
-
-		title := strings.TrimSpace(step.Title)
-		if title == "" {
-			title = step.ID
-		}
-
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: fmt.Sprintf("Executing step %s: %s", step.ID, title),
-			Level:   StatusLevelInfo,
-			Metadata: map[string]any{
-				"step_id": step.ID,
-				"title":   step.Title,
-				"command": step.Command.Run,
-				"shell":   step.Command.Shell,
-				"cwd":     step.Command.Cwd,
-			},
-		})
-
-		observation, err := r.executor.Execute(ctx, *step)
-		if ctx.Err() != nil {
-			finalErr = ctx.Err()
-			break
-		}
-
-		executedSteps++
-		lastStepID = step.ID
-
-		status := PlanCompleted
-		level := StatusLevelInfo
-		message := fmt.Sprintf("Step %s completed successfully.", step.ID)
-		if err != nil {
-			status = PlanFailed
-			level = StatusLevelError
-			if observation.Details == "" {
-				observation.Details = err.Error()
-			}
-			message = fmt.Sprintf("Step %s failed: %v", step.ID, err)
-			finalErr = err
-		}
-
-		result := StepObservation{
-			ID:        step.ID,
-			Status:    status,
-			Stdout:    observation.Stdout,
-			Stderr:    observation.Stderr,
-			ExitCode:  observation.ExitCode,
-			Details:   observation.Details,
-			Truncated: observation.Truncated,
-		}
-
-		planObservation := &PlanObservation{ObservationForLLM: &PlanObservationPayload{
-			PlanObservation: []StepObservation{result},
-		}}
-		if updateErr := r.plan.UpdateStatus(step.ID, status, planObservation); updateErr != nil {
-			r.emit(RuntimeEvent{
-				Type:    EventTypeError,
-				Message: fmt.Sprintf("Failed to update plan status for step %s: %v", step.ID, updateErr),
-				Level:   StatusLevelError,
-			})
-			finalErr = updateErr
-			break
-		}
-
-		lastObservation = observation
-		haveObservation = true
-		stepResults[step.ID] = result
-
-		metadata := map[string]any{
-			"step_id":   step.ID,
-			"title":     step.Title,
-			"status":    status,
-			"stdout":    observation.Stdout,
-			"stderr":    observation.Stderr,
-			"truncated": observation.Truncated,
-		}
-		if observation.ExitCode != nil {
-			metadata["exit_code"] = *observation.ExitCode
-		}
-		if observation.Details != "" {
-			metadata["details"] = observation.Details
-		}
-
-		r.emit(RuntimeEvent{
-			Type:     EventTypeStatus,
-			Message:  message,
-			Level:    level,
-			Metadata: metadata,
-		})
-
-		if err != nil {
-			break
-		}
-	}
-
-	orderedResults := make([]StepObservation, 0, len(stepResults))
-	if len(stepResults) > 0 {
-		ids := make([]string, 0, len(stepResults))
-		for id := range stepResults {
-			ids = append(ids, id)
-		}
-		sort.Strings(ids)
-		orderedResults = make([]StepObservation, 0, len(ids))
-		for _, id := range ids {
-			orderedResults = append(orderedResults, stepResults[id])
-		}
-	}
-
-	payload := PlanObservationPayload{PlanObservation: orderedResults}
-	if haveObservation {
-		payload.Stdout = lastObservation.Stdout
-		payload.Stderr = lastObservation.Stderr
-		payload.Truncated = lastObservation.Truncated
-		payload.ExitCode = lastObservation.ExitCode
-		payload.Details = lastObservation.Details
-	}
-
-	if payload.Summary == "" {
-		switch {
-		case executedSteps == 0 && finalErr != nil:
-			payload.Summary = "Failed before executing plan steps."
-		case executedSteps == 0:
-			payload.Summary = "No plan steps were executed."
-		case finalErr != nil:
-			payload.Summary = fmt.Sprintf("Execution halted during step %s.", lastStepID)
-		default:
-			payload.Summary = fmt.Sprintf("Executed %d plan step(s).", executedSteps)
-		}
-	}
-
-	r.appendToolObservation(toolCall, payload)
-}
-
-func (r *Runtime) appendToolObservation(toolCall ToolCall, payload PlanObservationPayload) {
-	if toolCall.ID == "" {
-		return
-	}
-
-	enforceObservationLimit(&payload)
-
-	toolMessage, err := BuildToolMessage(payload)
-	if err != nil {
-		r.emit(RuntimeEvent{
-			Type:    EventTypeError,
-			Message: fmt.Sprintf("Failed to encode tool observation: %v", err),
-			Level:   StatusLevelError,
-		})
-		return
-	}
-
-	r.appendHistory(ChatMessage{
-		Role:       RoleTool,
-		Content:    toolMessage,
-		ToolCallID: toolCall.ID,
-		Name:       toolCall.Name,
-		Timestamp:  time.Now(),
-	})
-}
-
-func (r *Runtime) beginWork() bool {
-	r.workMu.Lock()
-	defer r.workMu.Unlock()
-	if r.working {
-		return false
-	}
-	r.working = true
-	return true
-}
-
-func (r *Runtime) endWork() {
-	r.workMu.Lock()
-	r.working = false
-	r.workMu.Unlock()
-}
-
-func (r *Runtime) isWorking() bool {
-	r.workMu.Lock()
-	defer r.workMu.Unlock()
-	return r.working
-}
-
-func (r *Runtime) consumeInput(ctx context.Context) error {
-	scanner := bufio.NewScanner(r.options.InputReader)
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		default:
-		}
-
-		if !scanner.Scan() {
-			if err := scanner.Err(); err != nil {
-				return fmt.Errorf("failed to read input: %w", err)
-			}
-			r.Shutdown("stdin closed")
-			return nil
-		}
-
-		line := strings.TrimSpace(scanner.Text())
-		if r.isExitCommand(line) {
-			r.Shutdown("exit command received")
-			return nil
-		}
-
-		if strings.EqualFold(line, "cancel") {
-			r.Cancel("user requested cancel")
-			continue
-		}
-
-		r.SubmitPrompt(line)
-	}
-}
-
-func (r *Runtime) forwardOutputs(ctx context.Context) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case evt, ok := <-r.outputs:
-			if !ok {
-				return
-			}
-			fmt.Fprintf(r.options.OutputWriter, "[%s] %s\n", evt.Type, evt.Message)
-		}
-	}
-}
-
-func (r *Runtime) isExitCommand(value string) bool {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return false
-	}
-	for _, candidate := range r.options.ExitCommands {
-		if strings.EqualFold(trimmed, candidate) {
-			return true
-		}
-	}
-	return false
-}
-
 func (r *Runtime) emit(evt RuntimeEvent) {
 	select {
 	case <-r.closed:
@@ -1007,41 +148,6 @@ func (r *Runtime) close() {
 	})
 }
 
-func (r *Runtime) appendHistory(message ChatMessage) {
-	r.historyMu.Lock()
-	defer r.historyMu.Unlock()
-	r.history = append(r.history, message)
-}
-
-func (r *Runtime) historySnapshot() []ChatMessage {
-	r.historyMu.RLock()
-	defer r.historyMu.RUnlock()
-	copyHistory := make([]ChatMessage, len(r.history))
-	copy(copyHistory, r.history)
-	return copyHistory
-}
-
-func (r *Runtime) writeHistoryLog(history []ChatMessage) {
-	// Persist the exact payload forwarded to the model so hosts can inspect it.
-	data, err := json.MarshalIndent(history, "", "  ")
-	if err != nil {
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: fmt.Sprintf("Failed to encode history log: %v", err),
-			Level:   StatusLevelWarn,
-		})
-		return
-	}
-
-	if err := os.WriteFile("history.json", data, 0o644); err != nil {
-		r.emit(RuntimeEvent{
-			Type:    EventTypeStatus,
-			Message: fmt.Sprintf("Failed to write history log: %v", err),
-			Level:   StatusLevelWarn,
-		})
-	}
-}
-
 const baseSystemPrompt = `You are OpenAgent, an AI software engineer that plans and executes work.
 Always respond by calling the "open-agent" function tool with arguments that conform to the provided JSON schema.
 Explain your reasoning to the user in the "message" field and keep plans actionable, safe, and justified.
@@ -1056,8 +162,7 @@ If your task is to run a command, once you know that task is completed, to not r
 The plan is a Directed Acyclic Graph (DAG) of steps that can be executed in parallel when possible, do not assume order of independent steps.
 If order is required, use the "waitingForID" field to create dependencies between steps.
 Use the "requireHumanInput" field to pause execution and request additional input from the user when necessary.
-Be concise and clear in your reasoning and plan steps.
-`
+Be concise and clear in your reasoning and plan steps.`
 
 func buildSystemPrompt(augment string) string {
 	prompt := baseSystemPrompt
